+++ dataflow/flex-templates/hubspot-to-pubsub/src/main/java/com/akvelon/hubspot/templates/HubspotToPubSub.java
@@ -20,8 +20,8 @@
 package com.akvelon.hubspot.templates;
 
 import com.akvelon.hubspot.options.CdapHubspotOptions;
-import com.akvelon.hubspot.transforms.FormatInputTransform;
-import com.akvelon.hubspot.transforms.FormatOutputTransform;
+import com.akvelon.hubspot.transforms.HubspotFormatInput;
+import com.akvelon.hubspot.transforms.PubSubFormatOutput;
 import com.akvelon.hubspot.utils.JsonElementCoder;
 import com.akvelon.hubspot.utils.PluginConfigOptionsConverter;
 import com.google.gson.JsonElement;
@@ -41,7 +41,78 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * HubspotToPubsub Pipeline.
+ * The {@link HubspotToPubSub} Batch pipeline reading json encoded data from Hubspot and publishes
+ * to Google Cloud PubSub. Input topics, output topic. <br>
+ *
+ * <p><b>Pipeline Requirements</b>
+ *
+ * <ul>
+ *   <li>Hubspot private auth key.
+ *   <li>Hubspot Object(s) exists.
+ *   <li>The PubSub output topic exists.
+ * </ul>
+ *
+ * <p><b>Example Usage</b>
+ *
+ * <pre>
+ * # Set the pipeline vars
+ * PROJECT=id-of-my-project
+ * BUCKET_NAME=my-bucket
+ * REGION=my-region
+ *
+ * # Set containerization vars
+ * IMAGE_NAME=my-image-name
+ * TARGET_GCR_IMAGE=gcr.io/${PROJECT}/${IMAGE_NAME}
+ * BASE_CONTAINER_IMAGE=my-base-container-image
+ * BASE_CONTAINER_IMAGE_VERSION=my-base-container-image-version
+ * TEMPLATE_PATH="gs://${BUCKET_NAME}/templates/hubspot_to_pubsub.json"
+ * TARGET_GCR_IMAGE=gcr.io/${PROJECT}/${IMAGE_NAME}
+ *
+ * # Create bucket in the cloud storage
+ * gsutil mb gs://${BUCKET_NAME}
+ *
+ * # Go to the beam folder
+ * cd /path/to/flex-templates/hubspot-to-pubsub
+ *
+ * <b>FLEX TEMPLATE</b>
+ * # Assemble jar with dependencies
+ * mvn package -am -pl hubspot-to-pubsub
+ *
+ * # Go to the template folder
+ * cd /path/to/flex-templates/hubspot-to-pubsub
+ *
+ * # Build the flex template
+ * gcloud dataflow flex-template build ${TEMPLATE_PATH} \
+ *       --image-gcr-path "${TARGET_GCR_IMAGE}" \
+ *       --sdk-language "JAVA" \
+ *       --flex-template-base-image ${BASE_CONTAINER_IMAGE} \
+ *       --metadata-file "src/main/resources/hubspot_to_pubsub.json" \
+ *       --jar "target/hubspot-to-pubsub-1.0-SNAPSHOT.jar" \
+ *       --env FLEX_TEMPLATE_JAVA_MAIN_CLASS="com.akvelon.hubspot.templates.HubspotToPubSub"
+ *
+ * # Execute template:
+ * API_ROOT_URL="https://dataflow.googleapis.com"
+ * TEMPLATES_LAUNCH_API="${API_ROOT_URL}/v1b3/projects/${PROJECT}/locations/${REGION}/flexTemplates:launch"
+ * JOB_NAME="hubspot-to-pubsub-`date +%Y%m%d-%H%M%S-%N`"
+ *
+ * time curl -X POST -H "Content-Type: application/json" \
+ *         -H "Authorization: Bearer $(gcloud auth print-access-token)" \
+ *         -d '
+ *          {
+ *              "launch_parameter": {
+ *                  "jobName": "'$JOB_NAME'",
+ *                  "containerSpecGcsPath": "'$TEMPLATE_PATH'",
+ *                  "parameters": {
+ *                      "CDAP reference name": "Any String"
+ *                      "hubspot private application access token": "your-private-app-access-token",
+ *                      "hubspot objects to pull supported by CDAP": "Contacts",
+ *                      "outputTopic": "projects/'$PROJECT'/topics/your-topic-name"
+ *                  }
+ *              }
+ *          }
+ *         '
+ *         "${TEMPLATES_LAUNCH_API}"
+ * </pre>
  */
 public class HubspotToPubSub {
     /* Logger for class.*/
@@ -74,7 +145,7 @@ public class HubspotToPubSub {
         pipeline.getCoderRegistry().registerCoderForClass(JsonElement.class, JsonElementCoder.of());
 
         pipeline
-                .apply("readFromCdapHubspot", FormatInputTransform.readFromCdapHubspot(paramsMap))
+                .apply("readFromCdapHubspot", HubspotFormatInput.readFromCdapHubspot(paramsMap))
                 .setCoder(
                         KvCoder.of(
                                 NullableCoder.of(WritableCoder.of(NullWritable.class)), JsonElementCoder.of()))
@@ -91,7 +162,7 @@ public class HubspotToPubSub {
                         KvCoder.of(
                                 NullableCoder.of(WritableCoder.of(NullWritable.class)), StringUtf8Coder.of()))
                 .apply(Values.create())
-                .apply("writeToPubSub", new FormatOutputTransform.FormatOutput(options));
+                .apply("writeToPubSub", new PubSubFormatOutput(options));
 
         return pipeline.run();
     }
