+++ dataflow/flex-templates/pom.xml
@@ -119,7 +119,7 @@
         <dependency>
             <groupId>org.apache.beam</groupId>
             <artifactId>beam-sdks-java-io-sparkreceiver-2</artifactId>
-            <version>2.47.0</version>
+            <version>2.47.2</version>
             <exclusions>
                 <exclusion>
                     <groupId>org.slf4j</groupId>
